requirements.txt
fastapi
uvicorn
mangum

sentence-transformers
faiss-cpu